requirements.txt
fastapi
uvicorn
pillow
tensorflow
typing_extensions==4.7.1